+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * (behavior change) @EdgeType and @VertexType extensions now assign superclass directly to annotated type (or not if hierarchy already contains required type)
 * Call activateOnCurrentThread() in pools for each connection obtain (e.g. context.getConnection()) to guarantee proper db instance bound to thread
 * Fix guice circular proxy (between transaction manager and pools)
+* Fix child injector and private modules support for RepositoryModule (#14)
+* Fix playframework compatibility (#10)
 
 ### 3.1.1 (2015-08-16)
 * Improve graph connection recognition on repository methods: recognize all graph types implementing Vertex or Edge (e.g. OrientVertex, OrientEdge)